Add `LowRankLinear`
A linear layer applied to a sequence (NxC -> NxD) can be thought of as
having a weight tensor (NxDxC) where each slice along the sequence axis
is an exact copy. This is highly redundant. On the other hand, an untied
linear layer has fully independent weights along the sequence axis.

Intermediate between these extremes is to enforce the weight tensor to
be *low rank*, so that there is some independence but also some
redundancy. This is what we add here with `LowRankLinear`. We factorize
the weight tensor as the product of two weights (NxDxK, DxCxK). The
latent weight tensor is then computed as the product:

```
weight = einsum("ndk,dck->ndc", weight1, weight2)
```

If weight sharing is beneficial, the layer can easily "collapse" to a
rank one representation, while also having flexibility to learn some
diversity if it wants. It could be interesting to see if and when a
network chooses to use heterogeneous weights. Perhaps more diversity in
deeper layers?