Add simplified attention mechanisms
`Selection`: static feature based selection. Treats the input as both
key and value, and uses a static as opposed to dynamic query. The
hypothesis is that a static query may be sufficient when column weights
are untied. Intuitively, we think that the response selectivity of a
brain area does not depend on the input.

`Mixing`: static feature mixing only. This uses static mixing weights
similar to ConvNext and MLP-Mixer. Whereas in `Selection`, the query is
independent of input, here the attention weights are altogether
independent of input.