Add direct skip connections for recurrent models
Direct skip connections from earlier layers that bypass self-attention
are important in transformers. In the recurrent architecture however,
the skip input is from the same layer at the previous time step, not the
previous layer. Add an option to provide `direct_edges` to specify the
wiring of direct skip connections from previous layers.

Note, it is still the case that later layers do not receive direct input
until later time steps, relying instead only on self-attention for
communication. This could be something to explore and address later. But
in general, it seems somewhat bio plausible that later layers should not
respond strongly until they receive strong driving bottom up input. It
might be a bit of a waste of training compute though..

Also, did a bit of code cleanup, moved position embedding into the
`ColumnFormer`, and made `seq_len` required.