Apply direct edge input to attention
Previously I had "direct input" for feeding activations from lower
layers directly into the MLP, skipping attention. But this is not quite
the same as the classic feedforward ViT. What would really be consistent
is to have direct input into the attention. This way the q, k, v are
computed over the prior layer output. In addition though, you also get
k, v for the other layers.

Nb, it might also be worth trying a residual option, where the
self-recurrent input is added to the direct input.