Add initial wiring cost implementation
Each stage has an `in_geometry`, shape `(seq_len, in_seq_len)` and a
geometry, shape `(seq_len, seq_len)`. The stage is responsible for
computing the wiring cost on the appropriate pooling and attention
matrices. The losses get added to the output as a separate dict.

Add an `L1WiringCost` copied from `columnformers` and a
`CrossEntropyWiringCost`. Not sure if the cross-entropy one will be good
but just thought of it.