Debug block sparse linear
Could not use `sparse_bsr` layout weight parameter. It fails to map to
cuda correctly when calling `model.cuda()`. This is probably a bug that
should be reported. But as a workaround I just unpack the
`crow_indices`, `col_indices` as buffers and store the sparse bsr weight
values as a standard strided parameter. Then I construct the sparse bsr
weight tensor on the fly during forward.

TODO: backward does not work. Raises

```
RuntimeError: addmm: computation on CUDA is not implemented for Strided + Strided @ SparseBsr
```