Important bug fixes in quadformer/topomoe
Found two pretty critical bugs in forward impl for quadformer/topomoe
after initial experiments with quadformer, where I noticed the train
loss for the ViT equivalent wasn't going down as expected.

1. The residual path around attention was using normalized input rather
than raw input. This breaks the direct feedforward residual flow.

2. Cross attention was being applied for every block in the stage rather
than just the initial block. This could be a feature or bug. But
regardless it wasn't what I intended, and it wasn't even implemented
correctly.

After noticing these, I checked that the outputs of all my ViT
equivalents indeed match the reference timm ViT. This is a good check
that I should routinely do.

This also shows how easy it is for a tiny typo to slip past and cause a
subtle but significant issue.